Change main function call to call Allegro first.
Previously, Crystal owned the main function, then called Allegro's
al_run_main with a function pointer to a Crystal method, but this would
cause segfaults on Fiber switches.  Now, we make Allegro own the main
function, and pass Crystal's main function to it instead, which works as
expected.

We are keeping a slightly modified copy of redefine_main until
https://github.com/crystal-lang/crystal/pull/4798 is merged.